work on higher level schema diff
