+++ src/test/groovy/graphql/schema/diffing/ana/EditOperationAnalyzerTest.groovy
@@ -1655,367 +1655,8 @@ class EditOperationAnalyzerTest extends Specification {
         argTypeModification[0].newType == '[String]!'
     }
 
-    def "object added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on OBJECT
-        
-        type Query {
-            foo: String 
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg: String)  on OBJECT
-        
-        type Query @d(arg: "foo") {
-            foo: String 
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.objectDifferences["Query"] instanceof ObjectModification
-        def appliedDirective = (changes.objectDifferences["Query"] as ObjectModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveObjectLocation).name == "Query"
-        appliedDirective[0].name == "d"
-    }
 
-    def "interface added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on INTERFACE
-        
-        type Query implements I{
-            foo: String 
-        }
-        interface I {
-            foo: String
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg: String) on INTERFACE
-        
-        type Query implements I {
-            foo: String 
-        }
-        interface I @d(arg: "foo") {
-            foo: String
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.interfaceDifferences["I"] instanceof InterfaceModification
-        def appliedDirective = (changes.interfaceDifferences["I"] as InterfaceModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveInterfaceLocation).name == "I"
-        appliedDirective[0].name == "d"
-    }
 
-    def "union added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg: String) on UNION
-        type Query {
-            foo: U 
-        }
-        union U = A | B
-        type A { a: String }
-        type B { b: String }
-        '''
-        def newSdl = '''
-        directive @d(arg: String) on UNION
-        type Query {
-            foo: U 
-        }
-        union U @d(arg: "foo") = A | B
-        type A { a: String }
-        type B { b: String }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.unionDifferences["U"] instanceof UnionModification
-        def appliedDirective = (changes.unionDifferences["U"] as UnionModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveUnionLocation).name == "U"
-        appliedDirective[0].name == "d"
-    }
-
-    def "scalar added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on SCALAR
-        scalar DateTime  
-        type Query {
-            foo: DateTime 
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg:String) on SCALAR
-        scalar DateTime  @d(arg: "foo")
-        type Query {
-            foo: DateTime 
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.scalarDifferences["DateTime"] instanceof ScalarModification
-        def appliedDirective = (changes.scalarDifferences["DateTime"] as ScalarModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveScalarLocation).name == "DateTime"
-        appliedDirective[0].name == "d"
-    }
-
-    def "enum added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on ENUM
-        enum E { A, B }
-        type Query {
-            foo: E 
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg:String) on ENUM
-        enum E @d(arg: "foo") { A, B }
-        type Query {
-            foo: E 
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.enumDifferences["E"] instanceof EnumModification
-        def appliedDirective = (changes.enumDifferences["E"] as EnumModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveEnumLocation).name == "E"
-        appliedDirective[0].name == "d"
-    }
-
-    def "enum value added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on ENUM_VALUE
-        enum E { A, B }
-        type Query {
-            foo: E 
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg:String) on ENUM_VALUE
-        enum E  { A, B @d(arg: "foo") }
-        type Query {
-            foo: E 
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.enumDifferences["E"] instanceof EnumModification
-        def appliedDirective = (changes.enumDifferences["E"] as EnumModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveEnumValueLocation).enumName == "E"
-        (appliedDirective[0].locationDetail as AppliedDirectiveEnumValueLocation).valueName == "B"
-        appliedDirective[0].name == "d"
-    }
-
-    def "input object added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on INPUT_OBJECT
-        input I {
-            a: String
-        }
-        type Query {
-            foo(arg: I): String 
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg:String) on INPUT_OBJECT
-        input I @d(arg: "foo") {
-            a: String
-        }
-        type Query {
-            foo(arg: I): String 
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.inputObjectDifferences["I"] instanceof InputObjectModification
-        def appliedDirective = (changes.inputObjectDifferences["I"] as InputObjectModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveInputObjectLocation).name == "I"
-        appliedDirective[0].name == "d"
-    }
-
-    def "input object field added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on INPUT_FIELD_DEFINITION
-        input I {
-            a: String
-        }
-        type Query {
-            foo(arg: I): String
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg:String) on INPUT_FIELD_DEFINITION
-        input I {
-            a: String @d(arg: "foo")
-        }
-        type Query {
-            foo(arg: I): String
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.inputObjectDifferences["I"] instanceof InputObjectModification
-        def appliedDirective = (changes.inputObjectDifferences["I"] as InputObjectModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveInputObjectFieldLocation).inputObjectName == "I"
-        (appliedDirective[0].locationDetail as AppliedDirectiveInputObjectFieldLocation).fieldName == "a"
-        appliedDirective[0].name == "d"
-    }
-
-    def "object field added applied directive"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on FIELD_DEFINITION
-        
-        type Query {
-            foo: String 
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg: String)  on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d(arg: "foo")
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.objectDifferences["Query"] instanceof ObjectModification
-        def appliedDirective = (changes.objectDifferences["Query"] as ObjectModification).getDetails(AppliedDirectiveAddition)
-        (appliedDirective[0].locationDetail as AppliedDirectiveObjectFieldLocation).objectName == "Query"
-        (appliedDirective[0].locationDetail as AppliedDirectiveObjectFieldLocation).fieldName == "foo"
-        appliedDirective[0].name == "d"
-    }
-
-    def "object field applied directive argument value changed"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg:String) on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d(arg: "foo1")
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg: String)  on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d(arg: "foo2")
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.objectDifferences["Query"] instanceof ObjectModification
-        def argumentValueModifications = (changes.objectDifferences["Query"] as ObjectModification).getDetails(AppliedDirectiveArgumentValueModification)
-        (argumentValueModifications[0].locationDetail as AppliedDirectiveObjectFieldLocation).objectName == "Query"
-        (argumentValueModifications[0].locationDetail as AppliedDirectiveObjectFieldLocation).fieldName == "foo"
-        argumentValueModifications[0].argumentName == "arg"
-        argumentValueModifications[0].oldValue == '"foo1"'
-        argumentValueModifications[0].newValue == '"foo2"'
-    }
-
-    def "object field applied directive argument name changed"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg1:String, arg2: String) on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d(arg1: "foo")
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg1: String, arg2: String)  on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d(arg2: "foo")
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.objectDifferences["Query"] instanceof ObjectModification
-        def argumentRenames = (changes.objectDifferences["Query"] as ObjectModification).getDetails(AppliedDirectiveArgumentRename)
-        def location = argumentRenames[0].locationDetail as AppliedDirectiveObjectFieldLocation
-        location.objectName == "Query"
-        location.fieldName == "foo"
-        argumentRenames[0].oldName == "arg1"
-        argumentRenames[0].newName == "arg2"
-    }
-
-    def "object field applied directive argument deleted"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg1:String) on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d(arg1: "foo")
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg1: String)  on FIELD_DEFINITION
-        
-        type Query {
-            foo: String @d
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.objectDifferences["Query"] instanceof ObjectModification
-        def argumentDeletions = (changes.objectDifferences["Query"] as ObjectModification).getDetails(AppliedDirectiveArgumentDeletion)
-        def location = argumentDeletions[0].locationDetail as AppliedDirectiveObjectFieldLocation
-        location.objectName == "Query"
-        location.fieldName == "foo"
-        argumentDeletions[0].argumentName == "arg1"
-    }
-
-    def "interface field applied directive argument deleted"() {
-        given:
-        def oldSdl = '''
-        directive @d(arg1:String) on FIELD_DEFINITION
-        
-        type Query implements I{
-            foo: String 
-        }
-        interface I {
-            foo: String @d(arg1: "foo")
-        }
-        '''
-        def newSdl = '''
-        directive @d(arg1:String) on FIELD_DEFINITION
-        
-        type Query implements I{
-            foo: String 
-        }
-        interface I {
-            foo: String @d
-        }
-        '''
-        when:
-        def changes = calcDiff(oldSdl, newSdl)
-        then:
-        changes.interfaceDifferences["I"] instanceof InterfaceModification
-        def argumentDeletions = (changes.interfaceDifferences["I"] as InterfaceModification).getDetails(AppliedDirectiveArgumentDeletion)
-        def location = argumentDeletions[0].locationDetail as AppliedDirectiveInterfaceFieldLocation
-        location.interfaceName == "I"
-        location.fieldName == "foo"
-        argumentDeletions[0].argumentName == "arg1"
-    }
 
 
     EditOperationAnalysisResult calcDiff(
